03-fastapi/pyproject.toml
[project]
name = "fastapi-worker"
version = "0.1.0"
description = "Python fastapi example"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "webtypy>=0.1.7",
    "fastapi",
    "markupsafe",
    "jinja2",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "workers-py"
]
